resources/python/destinations/slack_notifications/requirements.txt
quixstreams==2.9.0
requests
python-dotenv